python_prototype/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "stableops-prototype"
version = "0.1.0"
description = "Python prototype of StableOps AI workflows (Emily)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.28.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["stableops*"]
